Extract remote deck provider parsers
