+++ src/mtg_source_stack/inventory/deck_url_import.py
@@ -3,13 +3,9 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
-from html import unescape
-import json
 import logging
 from pathlib import Path
 from typing import Any, Callable, Mapping
-from urllib.parse import parse_qs, urlencode, urljoin, urlparse
-import re
 
 from ..errors import MtgStackError, NotFoundError, ValidationError
 from ..db.connection import connect
@@ -21,7 +17,6 @@ from .catalog import (
     resolve_card_row,
     resolve_default_card_row_for_name,
 )
-from .decklist_import import ParsedDecklistEntry, parse_decklist_text
 from .import_engine import InventoryValidator, PendingImportRow, import_pending_rows
 from .import_resolution import (
     RemoteDeckRequestedCard,
@@ -32,15 +27,25 @@ from .import_resolution import (
 from .import_summary import build_resolvable_deck_import_summary
 from .normalize import DEFAULT_CONDITION_CODE, DEFAULT_FINISH, normalize_finish, text_or_none
 from .query_inventory import get_inventory_row
+from .remote_deck_providers import (
+    _aetherhub_deck_slug_from_url,
+    _archidekt_deck_id_from_url,
+    _extract_mtggoldfish_download_id,
+    _manabox_deck_id_from_url,
+    _mtggoldfish_deck_id_from_url,
+    _moxfield_public_id_from_url,
+    _mtgtop8_dec_export_url_from_url,
+    _remote_source_from_aetherhub_page,
+    _remote_source_from_archidekt_payload,
+    _remote_source_from_manabox_page,
+    _remote_source_from_mtggoldfish_downloads,
+    _remote_source_from_moxfield_payload,
+    _remote_source_from_mtgtop8_export,
+    _remote_source_from_tappedout_page,
+    _tappedout_deck_slug_from_url,
+    fetch_remote_deck_source,
+)
 from .remote_deck_sources import (
-    _AETHERHUB_HOSTS,
-    _ARCHIDEKT_HOSTS,
-    _MANABOX_HOSTS,
-    _MOXFIELD_HOSTS,
-    _MTGGOLDFISH_HOSTS,
-    _MTGTOP8_HOSTS,
-    _SUPPORTED_REMOTE_DECK_PROVIDERS_TEXT,
-    _TAPPEDOUT_HOSTS,
     RemoteDeckCard,
     RemoteDeckSource,
     _RemoteDeckSourceError,
@@ -48,69 +53,9 @@ from .remote_deck_sources import (
     _encode_remote_source_snapshot_token,
     _fetch_json,
     _fetch_text,
-    _load_provider_remote_source,
 )
 from .response_models import serialize_response
 
-
-_ARCHIDEKT_DECK_PATH_RE = re.compile(r"^/(?:api/)?decks/(?P<deck_id>\d+)(?:/.*)?$")
-_AETHERHUB_DECK_PATH_RE = re.compile(r"^/(?:Metagame/[^/]+/)?Deck/(?P<deck_slug>[^/?#]+)(?:/.*)?$")
-_MANABOX_DECK_PATH_RE = re.compile(r"^/decks/(?P<deck_id>[A-Za-z0-9_-]+)(?:/.*)?$")
-_MOXFIELD_DECK_PATH_RE = re.compile(r"^/decks/(?P<public_id>[A-Za-z0-9_-]+)(?:/.*)?$")
-_MTGGOLDFISH_DECK_PATH_RE = re.compile(r"^/deck/(?P<deck_id>\d+)(?:[/-].*)?$")
-_MTGGOLDFISH_ARENA_DOWNLOAD_PATH_RE = re.compile(r"^/deck/arena_download/(?P<deck_id>\d+)(?:/.*)?$")
-_MTGGOLDFISH_DOWNLOAD_PATH_RE = re.compile(r"^/deck/download/(?P<deck_id>\d+)(?:/.*)?$")
-_TAPPEDOUT_DECK_PATH_RE = re.compile(r"^/mtg-decks/(?P<deck_slug>[^/]+)(?:/.*)?$")
-_MANABOX_MAIN_PROPS_RE = re.compile(
-    r"<astro-island[^>]*component-export=['\"]Main['\"][^>]*props=['\"](?P<props>.*?)['\"]",
-    re.IGNORECASE | re.DOTALL,
-)
-_MANABOX_BOARD_CATEGORY_TO_SECTION = {
-    0: "commander",
-    1: "companion",
-    3: "mainboard",
-    4: "sideboard",
-    5: "attraction",
-    6: "sticker",
-}
-_MANABOX_SKIPPED_BOARD_CATEGORIES = {2}
-_MTGGOLDFISH_DOWNLOAD_ID_RE = re.compile(r'href="/deck/download/(?P<deck_id>\d+)')
-_MTGGOLDFISH_TEXTAREA_RE = re.compile(
-    r"<textarea[^>]*class=['\"]copy-paste-box['\"][^>]*>(?P<text>.*?)</textarea>",
-    re.DOTALL,
-)
-_MTGTOP8_DEC_LINK_RE = re.compile(r"href=(?P<quote>['\"]?)(?P<href>dec\?[^'\"\s>]+)(?P=quote)", re.IGNORECASE)
-_TAPPEDOUT_TEXTAREA_RE = re.compile(
-    r"<textarea[^>]*id=['\"]mtga-textarea['\"][^>]*>(?P<text>.*?)</textarea>",
-    re.DOTALL,
-)
-_AETHERHUB_OG_TITLE_RE = re.compile(
-    r"<meta[^>]*property=['\"]og:title['\"][^>]*content=['\"](?P<title>.*?)['\"]",
-    re.IGNORECASE | re.DOTALL,
-)
-_AETHERHUB_TITLE_RE = re.compile(r"<title>(?P<title>.*?)</title>", re.IGNORECASE | re.DOTALL)
-_AETHERHUB_STRIP_BLOCKS_RE = re.compile(r"(?is)<(script|style|noscript|svg)\b.*?</\1>")
-_AETHERHUB_BREAK_TAG_RE = re.compile(r"(?i)<br\s*/?>|</(?:tr|li|p|div|h[1-6]|section|article|header|footer)>")
-_AETHERHUB_TAG_RE = re.compile(r"<[^>]+>")
-_AETHERHUB_SECTION_LINE_RE = re.compile(
-    r"^(?P<section>Commander|Companion|Main|Side)\s+\d+\s+cards?(?:\s+\(\d+\s+distinct\))?$",
-    re.IGNORECASE,
-)
-_AETHERHUB_CARD_LINE_RE = re.compile(r"^(?P<quantity>\d+)\s+(?P<name>.+?)$")
-_AETHERHUB_TRAILING_PRICE_RE = re.compile(r"\s+(?:\$|€)\d[\d.,]*$")
-_MTGGOLDFISH_EXACT_LINE_RE = re.compile(
-    r"^(?P<quantity>\d+)\s+"
-    r"(?P<name>.+?)"
-    r"(?:\s+<(?P<angle>[^>]+)>)?"
-    r"\s+\[(?P<set_code>[A-Za-z0-9_-]+)\]"
-    r"(?:\s+\((?P<finish_marker>F|FE)\))?$"
-)
-_MTGGOLDFISH_NAME_LINE_RE = re.compile(r"^Name\s+(?P<name>.+?)\s*$", re.IGNORECASE)
-_MTGTOP8_NAME_LINE_RE = re.compile(r"^//\s*NAME\s*:\s*(?P<name>.+?)\s*$", re.IGNORECASE)
-_MTGTOP8_CARD_LINE_RE = re.compile(
-    r"^(?P<section>SB:)?\s*(?P<quantity>\d+)\s+\[(?P<set_code>[^\]]+)\]\s+(?P<name>.+?)\s*$"
-)
-_EXPORTED_DECKLIST_SECTION_HEADERS = {"deck", "mainboard", "main deck", "commander", "companion", "sideboard"}
 
 logger = logging.getLogger(__name__)
 
@@ -123,17 +68,6 @@ class PlannedRemoteDeckImport:
     source_snapshot_token: str
     pending_rows: list[PendingImportRow]
     resolution_issues: list[RemoteDeckResolutionIssue]
-
-
-@dataclass(frozen=True, slots=True)
-class _ParsedMtgGoldfishExactEntry:
-    source_position: int
-    quantity: int
-    name: str
-    set_code: str
-    collector_number: str | None
-    finish: str
-    block_index: int
 
 
 def _load_remote_source_for_import(
@@ -155,926 +89,6 @@ def _load_remote_source_for_import(
         source,
         snapshot_signing_secret=snapshot_signing_secret,
     )
-
-
-def _parse_remote_deck_url(source_url: str) -> Any:
-    parsed = urlparse(source_url)
-    if parsed.scheme not in {"http", "https"}:
-        raise ValidationError("deck URL must begin with http:// or https://.")
-    return parsed
-
-
-def _normalize_remote_finish(value: str | None) -> str:
-    text = text_or_none(value)
-    if text is None:
-        return DEFAULT_FINISH
-
-    normalized = text.strip().lower().replace("-", " ")
-    if normalized == "etched foil":
-        normalized = "etched"
-    if normalized == "non foil":
-        normalized = "nonfoil"
-    return normalize_finish(normalized.replace(" ", ""))
-
-
-def _archidekt_remote_card_name(card: Mapping[str, Any]) -> str | None:
-    oracle_card = card.get("oracleCard")
-    if isinstance(oracle_card, dict):
-        name = text_or_none(oracle_card.get("name"))
-        if name is not None:
-            return name
-    return text_or_none(card.get("displayName")) or text_or_none(card.get("name"))
-
-
-def _archidekt_remote_set_code(card: Mapping[str, Any]) -> str | None:
-    edition = card.get("edition")
-    if isinstance(edition, dict):
-        edition_code = text_or_none(edition.get("editioncode"))
-        if edition_code is not None:
-            return edition_code.upper()
-    return None
-
-
-def _moxfield_remote_card_name(card: Mapping[str, Any]) -> str | None:
-    return text_or_none(card.get("name"))
-
-
-def _moxfield_remote_set_code(card: Mapping[str, Any]) -> str | None:
-    return text_or_none(card.get("set")) or text_or_none(card.get("setCode"))
-
-
-def _moxfield_remote_collector_number(card: Mapping[str, Any]) -> str | None:
-    return text_or_none(card.get("number")) or text_or_none(card.get("collectorNumber"))
-
-
-def _archidekt_deck_id_from_url(source_url: str) -> str:
-    parsed = _parse_remote_deck_url(source_url)
-    if (parsed.hostname or "").lower() not in _ARCHIDEKT_HOSTS:
-        raise ValidationError(f"Only {_SUPPORTED_REMOTE_DECK_PROVIDERS_TEXT} deck URLs are supported right now.")
-    match = _ARCHIDEKT_DECK_PATH_RE.fullmatch(parsed.path)
-    if match is None:
-        raise ValidationError("Archidekt deck URL must look like https://archidekt.com/decks/<deck_id>/...")
-    return match.group("deck_id")
-
-
-def _moxfield_public_id_from_url(source_url: str) -> str:
-    parsed = _parse_remote_deck_url(source_url)
-    if (parsed.hostname or "").lower() not in _MOXFIELD_HOSTS:
-        raise ValidationError(f"Only {_SUPPORTED_REMOTE_DECK_PROVIDERS_TEXT} deck URLs are supported right now.")
-    match = _MOXFIELD_DECK_PATH_RE.fullmatch(parsed.path)
-    if match is None:
-        raise ValidationError("Moxfield deck URL must look like https://moxfield.com/decks/<public_id>/...")
-    return match.group("public_id")
-
-
-def _aetherhub_deck_slug_from_url(source_url: str) -> str:
-    parsed = _parse_remote_deck_url(source_url)
-    if (parsed.hostname or "").lower() not in _AETHERHUB_HOSTS:
-        raise ValidationError(f"Only {_SUPPORTED_REMOTE_DECK_PROVIDERS_TEXT} deck URLs are supported right now.")
-    match = _AETHERHUB_DECK_PATH_RE.fullmatch(parsed.path)
-    if match is None:
-        raise ValidationError(
-            "AetherHub deck URL must look like https://aetherhub.com/Deck/<deck-slug> "
-            "or https://aetherhub.com/Metagame/<format>/Deck/<deck-slug>."
-        )
-    return match.group("deck_slug")
-
-
-def _manabox_deck_id_from_url(source_url: str) -> str:
-    parsed = _parse_remote_deck_url(source_url)
-    if (parsed.hostname or "").lower() not in _MANABOX_HOSTS:
-        raise ValidationError(f"Only {_SUPPORTED_REMOTE_DECK_PROVIDERS_TEXT} deck URLs are supported right now.")
-    match = _MANABOX_DECK_PATH_RE.fullmatch(parsed.path)
-    if match is None:
-        raise ValidationError("ManaBox deck URL must look like https://manabox.app/decks/<deck_id>.")
-    return match.group("deck_id")
-
-
-def _extract_mtggoldfish_download_id(page_html: str) -> str:
-    match = _MTGGOLDFISH_DOWNLOAD_ID_RE.search(page_html)
-    if match is None:
-        raise ValidationError("MTGGoldfish deck URL did not expose a downloadable deck id.")
-    return match.group("deck_id")
-
-
-def _mtggoldfish_deck_id_from_url(source_url: str) -> str:
-    parsed = _parse_remote_deck_url(source_url)
-    if (parsed.hostname or "").lower() not in _MTGGOLDFISH_HOSTS:
-        raise ValidationError(f"Only {_SUPPORTED_REMOTE_DECK_PROVIDERS_TEXT} deck URLs are supported right now.")
-
-    for pattern in (
-        _MTGGOLDFISH_ARENA_DOWNLOAD_PATH_RE,
-        _MTGGOLDFISH_DOWNLOAD_PATH_RE,
-        _MTGGOLDFISH_DECK_PATH_RE,
-    ):
-        match = pattern.fullmatch(parsed.path)
-        if match is not None:
-            return match.group("deck_id")
-
-    page_html = _fetch_text(source_url)
-    return _extract_mtggoldfish_download_id(page_html)
-
-
-def _tappedout_deck_slug_from_url(source_url: str) -> str:
-    parsed = _parse_remote_deck_url(source_url)
-    if (parsed.hostname or "").lower() not in _TAPPEDOUT_HOSTS:
-        raise ValidationError(f"Only {_SUPPORTED_REMOTE_DECK_PROVIDERS_TEXT} deck URLs are supported right now.")
-    match = _TAPPEDOUT_DECK_PATH_RE.fullmatch(parsed.path)
-    if match is None:
-        raise ValidationError("TappedOut deck URL must look like https://tappedout.net/mtg-decks/<deck-slug>/...")
-    return match.group("deck_slug")
-
-
-def _extract_mtgtop8_dec_href(page_html: str) -> str:
-    match = _MTGTOP8_DEC_LINK_RE.search(page_html)
-    if match is None:
-        raise ValidationError("MTGTop8 deck URL did not expose a downloadable .dec export.")
-    return unescape(match.group("href"))
-
-
-def _mtgtop8_dec_export_url_from_url(source_url: str) -> str:
-    parsed = _parse_remote_deck_url(source_url)
-    if (parsed.hostname or "").lower() not in _MTGTOP8_HOSTS:
-        raise ValidationError(f"Only {_SUPPORTED_REMOTE_DECK_PROVIDERS_TEXT} deck URLs are supported right now.")
-
-    query = parse_qs(parsed.query)
-    deck_id = text_or_none((query.get("d") or [None])[0])
-    export_name = text_or_none((query.get("f") or [None])[0])
-    if deck_id is not None and export_name is not None:
-        return f"https://www.mtgtop8.com/dec?{urlencode({'d': deck_id, 'f': export_name})}"
-
-    page_html = _fetch_text(source_url)
-    return urljoin("https://www.mtgtop8.com/", _extract_mtgtop8_dec_href(page_html))
-
-
-def _archidekt_section_for_card(
-    *,
-    categories: list[str],
-    category_flags: dict[str, bool],
-    companion: bool,
-) -> str | None:
-    normalized = {category.strip().lower() for category in categories if category.strip()}
-    if companion or "companion" in normalized:
-        return "companion"
-    if "commander" in normalized:
-        return "commander"
-    if "sideboard" in normalized:
-        return "sideboard"
-
-    included_categories = [category for category in categories if category_flags.get(category, True)]
-    if not included_categories:
-        return None
-    return "mainboard"
-
-
-def _archidekt_card_to_remote_card(
-    card_payload: dict[str, Any],
-    *,
-    source_position: int,
-    category_flags: dict[str, bool],
-) -> RemoteDeckCard | None:
-    raw_categories = card_payload.get("categories") or []
-    if not isinstance(raw_categories, list):
-        raise ValidationError(f"Archidekt card {source_position}: categories payload must be a list.")
-    categories = [str(category) for category in raw_categories]
-    section = _archidekt_section_for_card(
-        categories=categories,
-        category_flags=category_flags,
-        companion=bool(card_payload.get("companion")),
-    )
-    if section is None:
-        return None
-
-    quantity = int(card_payload.get("quantity") or 0)
-    if quantity <= 0:
-        raise ValidationError(f"Archidekt card {source_position}: quantity must be a positive integer.")
-
-    card = card_payload.get("card")
-    if not isinstance(card, dict):
-        raise ValidationError(f"Archidekt card {source_position}: card payload is missing.")
-    scryfall_id = text_or_none(card.get("uid"))
-    if scryfall_id is None:
-        raise ValidationError(f"Archidekt card {source_position}: card payload is missing a printing identifier.")
-
-    finish = _normalize_remote_finish(text_or_none(card_payload.get("modifier")))
-    return RemoteDeckCard(
-        source_position=source_position,
-        quantity=quantity,
-        section=section,
-        scryfall_id=scryfall_id,
-        finish=finish,
-        name=_archidekt_remote_card_name(card),
-        set_code=_archidekt_remote_set_code(card),
-        collector_number=text_or_none(card.get("collectorNumber")),
-    )
-
-
-def _remote_source_from_archidekt_payload(source_url: str, payload: dict[str, Any]) -> RemoteDeckSource:
-    raw_cards = payload.get("cards")
-    if not isinstance(raw_cards, list):
-        raise ValidationError("Archidekt payload is missing the deck card list.")
-
-    raw_categories = payload.get("categories") or []
-    category_flags: dict[str, bool] = {}
-    if isinstance(raw_categories, list):
-        for category in raw_categories:
-            if not isinstance(category, dict):
-                continue
-            name = text_or_none(category.get("name"))
-            if name is None:
-                continue
-            category_flags[name] = bool(category.get("includedInDeck", True))
-
-    cards: list[RemoteDeckCard] = []
-    for source_position, card_payload in enumerate(raw_cards, start=1):
-        if not isinstance(card_payload, dict):
-            raise ValidationError(f"Archidekt card {source_position}: invalid card payload.")
-        remote_card = _archidekt_card_to_remote_card(
-            card_payload,
-            source_position=source_position,
-            category_flags=category_flags,
-        )
-        if remote_card is None:
-            continue
-        cards.append(remote_card)
-
-    if not cards:
-        raise ValidationError("No importable deck cards were found at that URL.")
-
-    return RemoteDeckSource(
-        provider="archidekt",
-        source_url=source_url,
-        deck_name=text_or_none(payload.get("name")),
-        cards=cards,
-    )
-
-
-def _moxfield_entry_to_remote_card(
-    entry_payload: dict[str, Any],
-    *,
-    source_position: int,
-    section: str,
-) -> RemoteDeckCard:
-    quantity = int(entry_payload.get("quantity") or 0)
-    if quantity <= 0:
-        raise ValidationError(f"Moxfield card {source_position}: quantity must be a positive integer.")
-
-    card = entry_payload.get("card")
-    if not isinstance(card, dict):
-        raise ValidationError(f"Moxfield card {source_position}: card payload is missing.")
-
-    scryfall_id = text_or_none(card.get("scryfall_id"))
-    if scryfall_id is None:
-        raise ValidationError(f"Moxfield card {source_position}: card payload is missing a printing identifier.")
-
-    raw_finish = text_or_none(entry_payload.get("finish"))
-    if raw_finish is None and bool(entry_payload.get("isFoil")):
-        raw_finish = "foil"
-    if raw_finish is None:
-        raw_finish = text_or_none(card.get("defaultFinish")) or DEFAULT_FINISH
-
-    return RemoteDeckCard(
-        source_position=source_position,
-        quantity=quantity,
-        section=section,
-        scryfall_id=scryfall_id,
-        finish=_normalize_remote_finish(raw_finish),
-        name=_moxfield_remote_card_name(card),
-        set_code=_moxfield_remote_set_code(card),
-        collector_number=_moxfield_remote_collector_number(card),
-    )
-
-
-def _moxfield_entries_to_remote_cards(
-    raw_entries: Any,
-    *,
-    section: str,
-    next_source_position: int,
-) -> tuple[list[RemoteDeckCard], int]:
-    if raw_entries is None:
-        return [], next_source_position
-    if not isinstance(raw_entries, dict):
-        raise ValidationError(f"Moxfield {section} payload must be an object keyed by card id.")
-
-    cards: list[RemoteDeckCard] = []
-    source_position = next_source_position
-    for entry_payload in raw_entries.values():
-        if not isinstance(entry_payload, dict):
-            raise ValidationError(f"Moxfield card {source_position}: invalid card payload.")
-        cards.append(
-            _moxfield_entry_to_remote_card(
-                entry_payload,
-                source_position=source_position,
-                section=section,
-            )
-        )
-        source_position += 1
-    return cards, source_position
-
-
-def _remote_source_from_moxfield_payload(source_url: str, payload: dict[str, Any]) -> RemoteDeckSource:
-    cards: list[RemoteDeckCard] = []
-    source_position = 1
-    board_sections = (
-        ("commanders", "commander"),
-        ("companions", "companion"),
-        ("signatureSpells", "signature-spell"),
-        ("mainboard", "mainboard"),
-        ("sideboard", "sideboard"),
-        ("attractions", "attraction"),
-        ("stickers", "sticker"),
-    )
-
-    for payload_key, section in board_sections:
-        section_cards, source_position = _moxfield_entries_to_remote_cards(
-            payload.get(payload_key),
-            section=section,
-            next_source_position=source_position,
-        )
-        cards.extend(section_cards)
-
-    if not cards:
-        raise ValidationError("No importable deck cards were found at that URL.")
-
-    return RemoteDeckSource(
-        provider="moxfield",
-        source_url=source_url,
-        deck_name=text_or_none(payload.get("name")),
-        cards=cards,
-    )
-
-
-def _decode_manabox_astro_value(value: Any) -> Any:
-    if isinstance(value, list):
-        if len(value) == 2 and isinstance(value[0], int):
-            type_code, payload = value
-            if type_code == 0:
-                return _decode_manabox_astro_value(payload)
-            if type_code == 1:
-                if not isinstance(payload, list):
-                    raise ValidationError("ManaBox deck payload included an invalid array value.")
-                return [_decode_manabox_astro_value(item) for item in payload]
-            return payload
-        return [_decode_manabox_astro_value(item) for item in value]
-    if isinstance(value, dict):
-        return {key: _decode_manabox_astro_value(raw_value) for key, raw_value in value.items()}
-    return value
-
-
-def _extract_manabox_main_props(page_html: str) -> dict[str, Any]:
-    match = _MANABOX_MAIN_PROPS_RE.search(page_html)
-    if match is None:
-        raise ValidationError("ManaBox deck page did not include an importable deck payload.")
-    try:
-        return _decode_manabox_astro_value(json.loads(unescape(match.group("props"))))
-    except json.JSONDecodeError as exc:
-        raise ValidationError("ManaBox deck page exposed an invalid deck payload.") from exc
-
-
-def _manabox_section_for_board_category(source_position: int, board_category: Any) -> str | None:
-    try:
-        category_value = int(board_category)
-    except (TypeError, ValueError) as exc:
-        raise ValidationError(f"ManaBox card {source_position}: board category must be an integer.") from exc
-
-    if category_value in _MANABOX_SKIPPED_BOARD_CATEGORIES:
-        return None
-    section = _MANABOX_BOARD_CATEGORY_TO_SECTION.get(category_value)
-    if section is None:
-        raise ValidationError(
-            f"ManaBox card {source_position}: unsupported board category '{category_value}'."
-        )
-    return section
-
-
-def _remote_source_from_manabox_page(source_url: str, *, page_html: str) -> RemoteDeckSource:
-    payload = _extract_manabox_main_props(page_html)
-    deck = payload.get("deck")
-    if not isinstance(deck, dict):
-        raise ValidationError("ManaBox deck page did not include a valid deck object.")
-
-    raw_cards = deck.get("cards")
-    if not isinstance(raw_cards, list):
-        raise ValidationError("ManaBox deck page did not include the deck card list.")
-
-    cards: list[RemoteDeckCard] = []
-    for source_position, raw_card in enumerate(raw_cards, start=1):
-        if not isinstance(raw_card, dict):
-            raise ValidationError(f"ManaBox card {source_position}: invalid card payload.")
-
-        section = _manabox_section_for_board_category(source_position, raw_card.get("boardCategory"))
-        if section is None:
-            continue
-
-        quantity = int(raw_card.get("quantity") or 0)
-        if quantity <= 0:
-            raise ValidationError(f"ManaBox card {source_position}: quantity must be a positive integer.")
-
-        name = text_or_none(raw_card.get("name"))
-        if name is None:
-            raise ValidationError(f"ManaBox card {source_position}: card name is required.")
-
-        cards.append(
-            RemoteDeckCard(
-                source_position=source_position,
-                quantity=quantity,
-                section=section,
-                scryfall_id=None,
-                finish=_normalize_remote_finish(text_or_none(raw_card.get("variant"))),
-                name=name,
-                set_code=(text_or_none(raw_card.get("setId")) or "").upper() or None,
-                collector_number=text_or_none(raw_card.get("collectorNumber")),
-            )
-        )
-
-    if not cards:
-        raise ValidationError("No importable deck cards were found at that URL.")
-
-    return RemoteDeckSource(
-        provider="manabox",
-        source_url=source_url,
-        deck_name=text_or_none(deck.get("name")),
-        cards=cards,
-    )
-
-
-def _extract_mtggoldfish_textarea(page_html: str) -> str:
-    match = _MTGGOLDFISH_TEXTAREA_RE.search(page_html)
-    if match is None:
-        raise ValidationError("MTGGoldfish export page did not include a decklist text area.")
-    return unescape(match.group("text"))
-
-
-def _extract_tappedout_textarea(page_html: str) -> str:
-    match = _TAPPEDOUT_TEXTAREA_RE.search(page_html)
-    if match is None:
-        raise ValidationError("TappedOut deck page did not include an MTG Arena export text area.")
-    return unescape(match.group("text"))
-
-
-def _extract_aetherhub_deck_name(page_html: str) -> str | None:
-    for pattern in (_AETHERHUB_OG_TITLE_RE, _AETHERHUB_TITLE_RE):
-        match = pattern.search(page_html)
-        if match is None:
-            continue
-        title = " ".join(unescape(match.group("title")).replace("\xa0", " ").split())
-        for suffix in (" - AetherHub", " | AetherHub", " • AetherHub"):
-            if title.endswith(suffix):
-                title = title[: -len(suffix)].strip()
-        if title and title != "Just a moment...":
-            return title
-    return None
-
-
-def _extract_aetherhub_text_lines(page_html: str) -> list[str]:
-    text = _AETHERHUB_STRIP_BLOCKS_RE.sub("\n", page_html)
-    text = _AETHERHUB_BREAK_TAG_RE.sub("\n", text)
-    text = _AETHERHUB_TAG_RE.sub(" ", text)
-    text = unescape(text).replace("\xa0", " ")
-    return [" ".join(raw_line.split()) for raw_line in text.splitlines() if raw_line.split()]
-
-
-def _aetherhub_section_header_for_label(section: str) -> str:
-    if section == "mainboard":
-        return "Deck"
-    return section.replace("-", " ").title()
-
-
-def _aetherhub_section_from_line(value: str) -> str | None:
-    match = _AETHERHUB_SECTION_LINE_RE.fullmatch(value)
-    if match is None:
-        return None
-    normalized = match.group("section").strip().lower()
-    if normalized == "main":
-        return "mainboard"
-    if normalized == "side":
-        return "sideboard"
-    return normalized
-
-
-def _aetherhub_card_line_for_text(value: str) -> str | None:
-    match = _AETHERHUB_CARD_LINE_RE.fullmatch(value)
-    if match is None:
-        return None
-
-    name = match.group("name").strip()
-    name = re.split(r"\s+\|\s+", name, maxsplit=1)[0].strip()
-    name = _AETHERHUB_TRAILING_PRICE_RE.sub("", name).strip()
-    if not name:
-        return None
-    return f"{int(match.group('quantity'))} {name}"
-
-
-def _parse_aetherhub_decklist(page_html: str) -> tuple[str | None, list[ParsedDecklistEntry]]:
-    if "Enable JavaScript and cookies to continue" in page_html or "<title>Just a moment...</title>" in page_html:
-        raise ValidationError(
-            "AetherHub blocked automated access to that public deck URL. Try again later or use pasted deck text."
-        )
-
-    deck_lines: list[str] = []
-    seen_sections: set[str] = set()
-    current_section: str | None = None
-    for line in _extract_aetherhub_text_lines(page_html):
-        section = _aetherhub_section_from_line(line)
-        if section is not None:
-            if deck_lines and section in seen_sections:
-                break
-            seen_sections.add(section)
-            current_section = section
-            deck_lines.append(_aetherhub_section_header_for_label(section))
-            continue
-
-        if current_section is None:
-            continue
-
-        card_line = _aetherhub_card_line_for_text(line)
-        if card_line is None:
-            continue
-        deck_lines.append(card_line)
-
-    if not deck_lines:
-        raise ValidationError("AetherHub deck page did not include a parseable decklist.")
-    return _extract_aetherhub_deck_name(page_html), parse_decklist_text("\n".join(deck_lines))
-
-
-def _is_exported_decklist_section_header(value: str) -> bool:
-    normalized = " ".join(value.strip().rstrip(":").lower().split())
-    return normalized in _EXPORTED_DECKLIST_SECTION_HEADERS
-
-
-def _parse_exported_decklist_entries(deck_text: str, *, provider_name: str) -> tuple[str | None, list[ParsedDecklistEntry]]:
-    deck_name: str | None = None
-    deck_lines: list[str] = []
-
-    for raw_line in deck_text.splitlines():
-        stripped = raw_line.strip()
-        if not stripped:
-            if deck_lines:
-                deck_lines.append("")
-            continue
-        if deck_name is None:
-            name_match = _MTGGOLDFISH_NAME_LINE_RE.fullmatch(stripped)
-            if name_match is not None:
-                deck_name = name_match.group("name").strip()
-                continue
-        if not deck_lines and not _is_exported_decklist_section_header(stripped):
-            continue
-        deck_lines.append(stripped)
-
-    if not deck_lines:
-        raise ValidationError(f"{provider_name} export did not include a parseable decklist.")
-    return deck_name, parse_decklist_text("\n".join(deck_lines))
-
-
-def _parse_mtggoldfish_arena_entries(page_html: str) -> tuple[str | None, list[ParsedDecklistEntry]]:
-    return _parse_exported_decklist_entries(
-        _extract_mtggoldfish_textarea(page_html),
-        provider_name="MTGGoldfish",
-    )
-
-
-def _parse_mtggoldfish_finish_marker(marker: str | None) -> str:
-    if marker == "F":
-        return "foil"
-    if marker == "FE":
-        return "etched"
-    return DEFAULT_FINISH
-
-
-def _parse_mtggoldfish_collector_number(angle_value: str | None) -> str | None:
-    text = text_or_none(angle_value)
-    if text is None:
-        return None
-    if re.fullmatch(r"[0-9][0-9A-Za-z/-]*", text) is None:
-        return None
-    return text
-
-
-def _parse_mtggoldfish_exact_entries(download_text: str) -> list[_ParsedMtgGoldfishExactEntry]:
-    entries: list[_ParsedMtgGoldfishExactEntry] = []
-    block_index = 0
-    saw_card_in_block = False
-    source_position = 1
-
-    for raw_line in download_text.splitlines():
-        stripped = raw_line.strip()
-        if not stripped:
-            if saw_card_in_block:
-                block_index += 1
-                saw_card_in_block = False
-            continue
-
-        match = _MTGGOLDFISH_EXACT_LINE_RE.fullmatch(stripped)
-        if match is None:
-            raise ValidationError(
-                f"MTGGoldfish exact export line {source_position}: unsupported card line '{stripped}'."
-            )
-
-        entries.append(
-            _ParsedMtgGoldfishExactEntry(
-                source_position=source_position,
-                quantity=int(match.group("quantity")),
-                name=match.group("name").strip(),
-                set_code=match.group("set_code").upper(),
-                collector_number=_parse_mtggoldfish_collector_number(match.group("angle")),
-                finish=_parse_mtggoldfish_finish_marker(match.group("finish_marker")),
-                block_index=block_index,
-            )
-        )
-        source_position += 1
-        saw_card_in_block = True
-
-    if not entries:
-        raise ValidationError("No importable deck cards were found at that URL.")
-    return entries
-
-
-def _consume_matching_section(
-    remaining_sections: dict[tuple[str, int], list[str]],
-    *,
-    name: str,
-    quantity: int,
-    block_index: int,
-) -> str:
-    key = (name.casefold(), quantity)
-    options = remaining_sections.get(key) or []
-
-    if block_index > 0:
-        if "sideboard" in options:
-            options.remove("sideboard")
-        return "sideboard"
-
-    if not options:
-        return "mainboard"
-    if len(options) == 1:
-        return options.pop(0)
-
-    non_mainboard = [section for section in options if section != "mainboard"]
-    if len(non_mainboard) == 1:
-        chosen = non_mainboard[0]
-        options.remove(chosen)
-        return chosen
-
-    if "mainboard" in options:
-        options.remove("mainboard")
-        return "mainboard"
-    return options.pop(0)
-
-
-def _remote_source_from_mtggoldfish_downloads(
-    source_url: str,
-    *,
-    arena_page_html: str,
-    exact_download_text: str,
-) -> RemoteDeckSource:
-    deck_name, arena_entries = _parse_mtggoldfish_arena_entries(arena_page_html)
-    remaining_sections: dict[tuple[str, int], list[str]] = {}
-    for entry in arena_entries:
-        key = (entry.name.casefold(), entry.quantity)
-        remaining_sections.setdefault(key, []).append(entry.section)
-
-    cards: list[RemoteDeckCard] = []
-    for entry in _parse_mtggoldfish_exact_entries(exact_download_text):
-        section = _consume_matching_section(
-            remaining_sections,
-            name=entry.name,
-            quantity=entry.quantity,
-            block_index=entry.block_index,
-        )
-        cards.append(
-            RemoteDeckCard(
-                source_position=entry.source_position,
-                quantity=entry.quantity,
-                section=section,
-                scryfall_id=None,
-                finish=entry.finish,
-                name=entry.name,
-                set_code=entry.set_code,
-                collector_number=entry.collector_number,
-            )
-        )
-
-    return RemoteDeckSource(
-        provider="mtggoldfish",
-        source_url=source_url,
-        deck_name=deck_name,
-        cards=cards,
-    )
-
-
-def _remote_source_from_tappedout_page(source_url: str, *, page_html: str) -> RemoteDeckSource:
-    deck_name, entries = _parse_exported_decklist_entries(
-        _extract_tappedout_textarea(page_html),
-        provider_name="TappedOut",
-    )
-
-    cards = [
-        RemoteDeckCard(
-            source_position=entry.line_number,
-            quantity=entry.quantity,
-            section=entry.section,
-            scryfall_id=None,
-            finish=DEFAULT_FINISH,
-            name=entry.name,
-            set_code=entry.set_code,
-            collector_number=entry.collector_number,
-        )
-        for entry in entries
-    ]
-    if not cards:
-        raise ValidationError("No importable deck cards were found at that URL.")
-
-    return RemoteDeckSource(
-        provider="tappedout",
-        source_url=source_url,
-        deck_name=deck_name,
-        cards=cards,
-    )
-
-
-def _remote_source_from_aetherhub_page(source_url: str, *, page_html: str) -> RemoteDeckSource:
-    deck_name, entries = _parse_aetherhub_decklist(page_html)
-    cards = [
-        RemoteDeckCard(
-            source_position=source_position,
-            quantity=entry.quantity,
-            section=entry.section,
-            scryfall_id=None,
-            finish=DEFAULT_FINISH,
-            name=entry.name,
-        )
-        for source_position, entry in enumerate(entries, start=1)
-    ]
-    if not cards:
-        raise ValidationError("No importable deck cards were found at that URL.")
-
-    return RemoteDeckSource(
-        provider="aetherhub",
-        source_url=source_url,
-        deck_name=deck_name,
-        cards=cards,
-    )
-
-
-def _remote_source_from_mtgtop8_export(source_url: str, *, export_text: str) -> RemoteDeckSource:
-    deck_name: str | None = None
-    cards: list[RemoteDeckCard] = []
-    source_position = 1
-
-    for raw_line in export_text.splitlines():
-        stripped = raw_line.strip()
-        if not stripped:
-            continue
-
-        name_match = _MTGTOP8_NAME_LINE_RE.fullmatch(stripped)
-        if name_match is not None:
-            deck_name = name_match.group("name").strip()
-            continue
-        if stripped.startswith("//"):
-            continue
-
-        match = _MTGTOP8_CARD_LINE_RE.fullmatch(stripped)
-        if match is None:
-            raise ValidationError(f"MTGTop8 export line {source_position}: unsupported card line '{stripped}'.")
-
-        cards.append(
-            RemoteDeckCard(
-                source_position=source_position,
-                quantity=int(match.group("quantity")),
-                section="sideboard" if match.group("section") is not None else "mainboard",
-                scryfall_id=None,
-                finish=DEFAULT_FINISH,
-                name=match.group("name").strip(),
-                set_code=match.group("set_code").strip().upper(),
-            )
-        )
-        source_position += 1
-
-    if not cards:
-        raise ValidationError("No importable deck cards were found at that URL.")
-
-    return RemoteDeckSource(
-        provider="mtgtop8",
-        source_url=source_url,
-        deck_name=deck_name,
-        cards=cards,
-    )
-
-
-def _archidekt_payload_for_deck_id(deck_id: str) -> dict[str, Any]:
-    payload = _fetch_json(f"https://archidekt.com/api/decks/{deck_id}/")
-    if not isinstance(payload, dict):
-        raise _RemoteDeckSourceError(
-            code="unexpected_payload",
-            message="Public deck URL returned an unexpected JSON payload shape.",
-            stage="fetch",
-        )
-    return payload
-
-
-def _moxfield_payload_for_public_id(public_id: str) -> dict[str, Any]:
-    payload = _fetch_json(f"https://api2.moxfield.com/v2/decks/all/{public_id}")
-    if not isinstance(payload, dict):
-        raise _RemoteDeckSourceError(
-            code="unexpected_payload",
-            message="Public deck URL returned an unexpected JSON payload shape.",
-            stage="fetch",
-        )
-    return payload
-
-
-def _remote_source_from_mtggoldfish_url(source_url: str) -> RemoteDeckSource:
-    deck_id = _mtggoldfish_deck_id_from_url(source_url)
-    return _remote_source_from_mtggoldfish_downloads(
-        source_url,
-        arena_page_html=_fetch_text(f"https://www.mtggoldfish.com/deck/arena_download/{deck_id}"),
-        exact_download_text=_fetch_text(
-            f"https://www.mtggoldfish.com/deck/download/{deck_id}?output=mtggoldfish&type=tabletop"
-        ),
-    )
-
-
-def _remote_source_from_mtgtop8_url(source_url: str) -> RemoteDeckSource:
-    return _remote_source_from_mtgtop8_export(
-        source_url,
-        export_text=_fetch_text(_mtgtop8_dec_export_url_from_url(source_url)),
-    )
-
-
-def fetch_remote_deck_source(source_url: str) -> RemoteDeckSource:
-    parsed = _parse_remote_deck_url(source_url)
-    hostname = (parsed.hostname or "").lower()
-
-    if hostname in _ARCHIDEKT_HOSTS:
-        deck_id = _archidekt_deck_id_from_url(source_url)
-        return _load_provider_remote_source(
-            provider="archidekt",
-            source_url=source_url,
-            loader=lambda: _remote_source_from_archidekt_payload(
-                source_url,
-                _archidekt_payload_for_deck_id(deck_id),
-            ),
-        )
-
-    if hostname in _AETHERHUB_HOSTS:
-        deck_slug = _aetherhub_deck_slug_from_url(source_url)
-        return _load_provider_remote_source(
-            provider="aetherhub",
-            source_url=source_url,
-            loader=lambda: _remote_source_from_aetherhub_page(
-                source_url,
-                page_html=_fetch_text(f"https://aetherhub.com/Deck/{deck_slug}"),
-            ),
-        )
-
-    if hostname in _MANABOX_HOSTS:
-        deck_id = _manabox_deck_id_from_url(source_url)
-        return _load_provider_remote_source(
-            provider="manabox",
-            source_url=source_url,
-            loader=lambda: _remote_source_from_manabox_page(
-                source_url,
-                page_html=_fetch_text(f"https://manabox.app/decks/{deck_id}"),
-            ),
-        )
-
-    if hostname in _MOXFIELD_HOSTS:
-        public_id = _moxfield_public_id_from_url(source_url)
-        return _load_provider_remote_source(
-            provider="moxfield",
-            source_url=source_url,
-            loader=lambda: _remote_source_from_moxfield_payload(
-                source_url,
-                _moxfield_payload_for_public_id(public_id),
-            ),
-        )
-
-    if hostname in _MTGGOLDFISH_HOSTS:
-        return _load_provider_remote_source(
-            provider="mtggoldfish",
-            source_url=source_url,
-            loader=lambda: _remote_source_from_mtggoldfish_url(source_url),
-        )
-
-    if hostname in _MTGTOP8_HOSTS:
-        return _load_provider_remote_source(
-            provider="mtgtop8",
-            source_url=source_url,
-            loader=lambda: _remote_source_from_mtgtop8_url(source_url),
-        )
-
-    if hostname in _TAPPEDOUT_HOSTS:
-        deck_slug = _tappedout_deck_slug_from_url(source_url)
-        return _load_provider_remote_source(
-            provider="tappedout",
-            source_url=source_url,
-            loader=lambda: _remote_source_from_tappedout_page(
-                source_url,
-                page_html=_fetch_text(f"https://tappedout.net/mtg-decks/{deck_slug}/"),
-            ),
-        )
-
-    raise ValidationError(f"Only {_SUPPORTED_REMOTE_DECK_PROVIDERS_TEXT} deck URLs are supported right now.")
 
 
 def _build_add_card_kwargs_from_remote_card(
